record_syscall: Store RDMA_VERBS_IOCTL attrs in scratch
`struct ib_uverbs_ioctl_hdr` as an VLA of `struct ib_uverbs_attr`. Each of this
can hold a pointer to userspace buffer or a piece of data.

Kernel driver distinguish between data/pointer based on internal kernel data,
so here memory buffer is recorded in scratch if:

- data length is longer than 8 (then it is most certainly a buffer);
- or data length is equal to 8 and data addr points to a valid tracee memory
  (however it might not be accurate, data length might be 8 and can be treated as data,
  but store scratch anyway).

On top of that attr struct is always stored in scratch.